+++ components/src/core/header/headerIcon.tsx
@@ -1,11 +1,12 @@
-import React, { ComponentType } from 'react';
-import { WrapIconProps } from '../icon-wrapper';
+import React from 'react';
+import { Icon } from '../icon';
+import { IconSource } from '../__types__';
 import { ICON_SIZE } from './constants';
 import { useColor } from './contexts/ColorContextProvider';
 
 type HeaderIconProps = {
     /** A component to render for the icon  */
-    IconClass?: ComponentType<WrapIconProps>;
+    icon?: IconSource;
 };
 /**
  * HeaderIcon component
@@ -13,10 +14,10 @@ type HeaderIconProps = {
  * The HeaderIcon is a helper component that is used to properly size and space icons in the Header component.
  */
 export const HeaderIcon: React.FC<HeaderIconProps> = (props) => {
-    const { IconClass } = props;
+    const { icon } = props;
     const { color } = useColor();
-    if (IconClass) {
-        return <IconClass size={ICON_SIZE} color={color} allowFontScaling />;
+    if (icon) {
+        return <Icon source={icon} size={ICON_SIZE} color={color} allowFontScaling />;
     }
     return null;
 };