+++ demos/storybook/storybook/stories/info-list-item/info-list-item.tsx
@@ -1,25 +1,21 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react-native';
-import { InfoListItem, ChannelValue, wrapIcon } from '@pxblue/react-native-components';
+import { InfoListItem, ChannelValue } from '@pxblue/react-native-components';
 import { text, boolean, withKnobs, color } from '@storybook/addon-knobs';
-import Leaf from '@pxblue/icons-svg/leaf.svg';
-import Temp from '@pxblue/icons-svg/temp.svg';
-import A from '@pxblue/icons-svg/grade_a.svg';
-import Device from '@pxblue/icons-svg/device.svg';
 import { framedRow } from '../../decorators';
 import * as Colors from '@pxblue/colors';
-import MatIcon from 'react-native-vector-icons/MaterialIcons';
+import { IconFamily } from '@pxblue/react-native-components/core/__types__';
 
 const notes = {
     notes:
         'The borders are NOT part of the component; they are provided for framing only. Any React Element may be passed in as `icon`; if using an svg, its color and size are not controlled by `ChannelValue`',
 };
 
-const LeafIcon = wrapIcon({ IconClass: Leaf });
-const AlarmIcon = wrapIcon({ IconClass: MatIcon, name: 'alarm' });
-const TempIcon = wrapIcon({ IconClass: Temp });
-const DeviceIcon = wrapIcon({ IconClass: Device });
-const AIcon = wrapIcon({ IconClass: A });
+const LeafIcon: IconFamily = { family: 'pxblue', name: 'leaf' };
+const AlarmIcon: IconFamily = { name: 'alarm' };
+const TempIcon: IconFamily = { family: 'pxblue', name: 'temp' };
+const DeviceIcon: IconFamily = { family: 'pxblue', name: 'device' };
+const AIcon: IconFamily = { family: 'pxblue', name: 'grade_a' };
 storiesOf('InfoListItem', module)
     .addDecorator(withKnobs)
     .addDecorator(framedRow)
@@ -41,7 +37,7 @@ storiesOf('InfoListItem', module)
                 title={'Info List Item'}
                 // TODO: make this work
                 // iconAlign={select('iconAlign', ['left', 'center', 'right'], 'left')}
-                IconClass={AlarmIcon}
+                icon={AlarmIcon}
                 iconColor={color('iconColor', Colors.black[500])}
                 subtitle={'with an icon'}
             />
@@ -53,7 +49,7 @@ storiesOf('InfoListItem', module)
         () => (
             <InfoListItem
                 title={'Info List Item'}
-                IconClass={TempIcon}
+                icon={TempIcon}
                 subtitle={[
                     <ChannelValue key={'temp1'} value={'50'} units={'°C'} />,
                     <ChannelValue key={'temp2'} value={'55'} units={'°C'} />,
@@ -75,7 +71,7 @@ storiesOf('InfoListItem', module)
         () => (
             <InfoListItem
                 title={'Info List Item'}
-                IconClass={AIcon}
+                icon={AIcon}
                 avatar={boolean('avatar', true)}
                 subtitle={'with an avatar and configurable status color'}
                 statusColor={color('statusColor', Colors.green[700])}
@@ -88,7 +84,7 @@ storiesOf('InfoListItem', module)
         () => (
             <InfoListItem
                 title={'Info List Item'}
-                IconClass={LeafIcon}
+                icon={LeafIcon}
                 iconColor={Colors.black[500]}
                 subtitle={'With a configurable background color'}
                 fontColor={Colors.black[500]}
@@ -102,7 +98,7 @@ storiesOf('InfoListItem', module)
         () => (
             <InfoListItem
                 title={'Info List Item'}
-                IconClass={DeviceIcon}
+                icon={DeviceIcon}
                 subtitle={'with a right component'}
                 rightComponent={
                     <ChannelValue
@@ -119,7 +115,7 @@ storiesOf('InfoListItem', module)
         () => (
             <InfoListItem
                 title={text('title', 'Info List Item')}
-                IconClass={DeviceIcon}
+                icon={DeviceIcon}
                 subtitle={text('subtitle', 'with all customizable properties')}
                 info={text('info', 'more info...')}
                 onPress={